adding main page
